Cleanup and tidying
