+++ components/led_panel/led_panel_test.py
@@ -46,7 +46,6 @@ class LedPanelTest(object):
                     time.sleep(1)
                     self.__panel.display(self.__matrix_right)
                 elif cmd_opts.get("bob"):
-                    # self.__panel.display(self.__bob_face)
                     self.__panel.display(self.__bob_eyes_open)
                     time.sleep(1)
                     self.__panel.display(self.__bob_eyes_shut)
